library_data_analysis_fastapi/pyproject.toml
[tool.poetry]
name = "library-data-analysis"
version = "0.10.0"
description = "图书馆数据分析系统后端"
authors = ["Library Data Analysis Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.136.0"
uvicorn = "^0.34.0"
psycopg2-binary = "^2.9.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
bcrypt = "^4.0.0"
Pillow = "^12.0.0"
python-multipart = "^0.0.20"
pyjwt = "^2.10.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
alembic = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
